changes: * remove service wrappers for imported services * remove unused code * exclude domains manager to separate component
