+++ apps/web/src/app/status-pages/components/create-edit-form.tsx
@@ -1,4 +1,3 @@
-import { useState } from "react";
 import { Button } from "@/components/ui/button";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Input } from "@/components/ui/input";
@@ -18,7 +17,8 @@ import { z } from "zod";
 import { useForm } from "react-hook-form";
 import { useLocalizedTranslation } from "@/hooks/useTranslation";
 import { Alert, AlertDescription } from "@/components/ui/alert";
-import { AlertCircleIcon, PlusIcon, XIcon } from "lucide-react";
+import { AlertCircleIcon } from "lucide-react";
+import DomainsManager from "./domains-manager";
 
 const statusPageSchema = z.object({
   title: z.string().min(1, "Title is required"),
@@ -59,82 +59,7 @@ const formDefaultValues: StatusPageForm = {
   domains: [],
 };
 
-const DomainsManager = ({
-  value = [],
-  onChange,
-}: {
-  value?: string[];
-  onChange: (domains: string[]) => void;
-}) => {
-  const { t } = useLocalizedTranslation();
-  const [newDomain, setNewDomain] = useState("");
 
-  const addDomain = () => {
-    if (newDomain.trim() && !value.includes(newDomain.trim())) {
-      onChange([...value, newDomain.trim()]);
-      setNewDomain("");
-    }
-  };
-
-  const removeDomain = (index: number) => {
-    const updated = value.filter((_, i) => i !== index);
-    onChange(updated);
-  };
-
-  const handleKeyPress = (e: React.KeyboardEvent) => {
-    if (e.key === "Enter") {
-      e.preventDefault();
-      addDomain();
-    }
-  };
-
-  return (
-    <div className="space-y-3">
-      {value.length > 0 && (
-        <div className="space-y-2">
-          <p className="text-sm font-medium">{t("forms.placeholders.domains")}</p>
-          <div className="space-y-1">
-            {value.map((domain, index) => (
-              <div
-                key={index}
-                className="flex items-center justify-between bg-muted p-2 rounded-md"
-              >
-                <span className="text-sm">{domain}</span>
-                <Button
-                  type="button"
-                  variant="ghost"
-                  size="sm"
-                  onClick={() => removeDomain(index)}
-                  className="h-6 w-6 p-0"
-                >
-                  <XIcon className="h-3 w-3" />
-                </Button>
-              </div>
-            ))}
-          </div>
-        </div>
-      )}
-
-      <div className="flex gap-2">
-        <Input
-          placeholder={t("forms.placeholders.domains")}
-          value={newDomain}
-          onChange={(e) => setNewDomain(e.target.value)}
-          onKeyDown={handleKeyPress}
-          className="flex-1"
-        />
-        <Button
-          type="button"
-          onClick={addDomain}
-          disabled={!newDomain.trim() || value.includes(newDomain.trim())}
-          size="sm"
-        >
-          <PlusIcon className="h-4 w-4" />
-        </Button>
-      </div>
-    </div>
-  );
-};
 
 const CreateEditForm = ({
   onSubmit,
